+++ pytest_bdd/types.py
@@ -3,10 +3,11 @@
 FEATURE = "feature"
 SCENARIO_OUTLINE = "scenario outline"
 EXAMPLES = "examples"
-EXAMPLES_VERTICAL = "examples vertical"
+EXAMPLES_VERTICAL_LEGACY = "examples vertical"  # Deprecated
+EXAMPLES_TRANSPOSED = "examples transposed"
 EXAMPLES_HEADERS = "example headers"
 EXAMPLE_LINE = "example line"
-EXAMPLE_LINE_VERTICAL = "example line vertical"
+EXAMPLE_LINE_TRANSPOSED = "example line transposed"
 SCENARIO = "scenario"
 BACKGROUND = "background"
 GIVEN = "given"